fix(deep_causality_discovery): refactored CDL module for improved modularity
